refactor: move workflow initialization calls to the <Builder />
